Removed warnings reported by gcc.

+++ src/timeconv.inl
@@ -38,7 +38,7 @@ namespace {
     {
         ts.tv_sec = static_cast<int>(xt.sec);
         ts.tv_nsec = static_cast<int>(xt.nsec);
-        if(ts.tv_nsec > NANOSECONDS_PER_SECOND)
+        if(ts.tv_nsec > static_cast<const int>(NANOSECONDS_PER_SECOND))
         {
             ts.tv_sec += ts.tv_nsec / NANOSECONDS_PER_SECOND;
             ts.tv_nsec %= NANOSECONDS_PER_SECOND;
@@ -74,7 +74,7 @@ namespace {
                 ts.tv_sec -= 1;
                 ts.tv_nsec += NANOSECONDS_PER_SECOND;
             }
-           if(ts.tv_nsec > NANOSECONDS_PER_SECOND)
+           if(ts.tv_nsec > static_cast<const int>(NANOSECONDS_PER_SECOND))
            {
                ts.tv_sec += ts.tv_nsec / NANOSECONDS_PER_SECOND;
                ts.tv_nsec %= NANOSECONDS_PER_SECOND;
